// src/app/_core/models/compte-facturation-vo.ts
import { EntityBase } from './entity-base';
import { MouvementECVO } from './mouvementECVO';
import { DossierCompteFacturationVO } from './dossier-compte-facturation-vo';
import { InvoiceVO } from './invoice-vo';



export interface CompteFacturation extends EntityBase {

 identifier: string;
 customerIdentifier: string;
 compteRenduAttendu: Boolean;
 universe: string;
 alias: string;
 status: string;
 nicheAdmissionDate: string;
 terminationDate: string;
 paymentMode: string;
 paymentCondition: string;
 accountName: string;
 bankCode: string;
 bankName: string;
 officeCode: string;
 officeName: string;
 bankAccountNumber: string;
 ribKey: string;
 payerTitle: string;
 payerLastName: string;
 payerFirstName: string;
 payerCompany: string;
 payerSiren: string;
 payerAddress2: string;
 payerAddress3: string;
 payerAddress4: string;
 payerAddress5: string;
 payerAddress6: string;
 payerDeliveryInfo: string;
 payerStreetNumber: Number;
 payerStreetExtension: string;
 payerStreetType: string;
 payerStreetName: string;
 payerOrasId: Number;
 payerInseeId: string;
 payerRivoliId: string;
 payerGeocodeX: Number;
 payerGeocodeY: Number;
 payerZipCode: string;
 payerCity: string;
 payerCedex: string;
 payerCountry: string;
 titularTitle: string;
 titularLastName: string;
 titularFirstName: string;
 titularCompany: string;
 titularSiren: string;
 titularAddress2: string;
 titularAddress3: string;
 titularAddress4: string;
 titularAddress5: string;
 titularAddress6: string;
 titularDeliveryInfo: string;
 titularStreetNumber: Number;
 titularStreetExtension: string;
 titularStreetType: string;
 titularStreetName: string;
 titularOrasId: Number;
 titularInseeId: string;
 titularRivoliId: string;
 titularGeocodeX: Number;
 titularGeocodeY: Number;
 titularZipCode: string;
 titularCity: string;
 titularCedex: string;
 titularCountry: string;
 recurrent: Boolean;
 balance: Number;
 representedAmount: Number;
 billingDay: Number;
 timeBeforeBilling: Number;
 mailNotification: string;
 controlCF: Boolean;
 justificationControle: string;
 tvaExemption: Boolean;
 billAccountLines: Array<DossierCompteFacturationVO>;
 allTypeEnvoi: any;
 isPluriel: Boolean;
 isPro: Boolean;
 ibanCode: string;
 mouvementsCf: Array<MouvementECVO>;
 modeReglementLibelle: string;
 conditionReglementLibelle: string;
 dateMaj : string;
 dateCreation : string;
 idCompteFacturation : Number;
 invoices : Array<InvoiceVO>;
}